requirements.txt
Flask>=1.1.2
pymongo>=3.11.4
python-dotenv>=0.17.0
stripe>=7.0.0
urllib3>=1.26.7
cryptography>=3.4.7